requirements.txt
# Data Processing & Analysis
numpy==1.24.3
pandas==2.0.3
plotly==5.18.0

# HTTP & API
requests==2.31.0
flask==2.0.1
flask-cors==4.0.0

# Database
sqlalchemy==2.0.23

# Configuration & Environment
python-dotenv==1.0.0

# Task Scheduling
schedule==1.2.0

# Testing
pytest==7.4.2